requirements.txt
Flask==1.1.2
requests==2.25.0
gunicorn==20.0.4
Pillow
pypng
pyqrcode
numpy
textwrap